Update TypeScript & build, add publish to gh-pages
Updating TypeScript broken our build, so dance like a monkey to get
that all working again.  Fixing the current build was easy enough, fix
a compilation error and change the default tsconfig.json so we do not
have to recompile for npm publication.

Adding automatic publication of TypeDocs was a little more
complicated.  TypeDoc does not like the new default tsconfig.json
format (it has comments), so revert that, but simplify tsconfig.json
to take advantage of defaults and be strict.  Also, typedoc is run
with `--mode modules` and a tsconfig.json file exists, it reads the
tsconfig.json and ignores any file arguments and --exclude options on
the command line and produces TypeDocs with the src and test
subdirectories in them.  The best solution I could come up with was to
generate a TypeDoc specific tsconfig.json and use that when running
typedoc.

Improve docs and add link to the TypeDoc on gh-pages.

Move various build scripts to scripts directory.